stuffs
+ Room colliders need fixing (Tilemap is messing up all the collider)

+++ Assets/Scripts/Objects/Enemy/Enemy 2.0/EnemyHealthUI.cs
@@ -7,14 +7,14 @@ public class EnemyHealthUI : MonoBehaviour
 {
     [SerializeField] private Slider slider;
     
-    public void SetMaxHealth(int maxHealth)
+    public void SetMaxHealth(float maxHealth)
     {
         slider.maxValue = maxHealth;
         slider.value = maxHealth; 
     }
 
 
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         slider.value = health;
     }
